refactor: extract pure summary calculations into _lib/calculations
Mechanical move of calculateAccountFinalBalance and calculateMonthlySummary
out of the "use server" module so they can be unit-tested. No behavior change.
